Implementation of the suggestions presented by the reviewer.
